+++ src/main/java/eu/gir/girsignals/blocks/SignalBlock.java
@@ -37,6 +37,8 @@ import net.minecraft.world.World;
 import net.minecraftforge.common.property.ExtendedBlockState;
 import net.minecraftforge.common.property.IExtendedBlockState;
 import net.minecraftforge.common.property.IUnlistedProperty;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class SignalBlock extends Block implements ITileEntityProvider {
 
@@ -261,4 +263,19 @@ public class SignalBlock extends Block implements ITileEntityProvider {
 	public String toString() {
 		return this.getLocalizedName();
 	}
+	
+	public boolean canBeLinked() {
+		return true;
+	}
+	
+	@SideOnly(Side.CLIENT)
+	public int colorMultiplier(IBlockState state, IBlockAccess worldIn, BlockPos pos, int tintIndex) {
+		return 0;
+	}
+	
+	@SideOnly(Side.CLIENT)
+	public boolean hasCostumColor() {
+		return false;
+	}
+	
 }